pyproject.toml
[project]
name = "mcp_pazy"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "iagopdiaz", email = "iago.pdiaz@udc.es"},
    {name = "IagoSuntzu", email = "iago@suntzu-growth.com"},
]
dependencies = [
    "langgraph>=1.0.8",
    "langchain>=1.2.9",
    "pandas>=3.0.0",
    "openpyxl>=3.1.5",
    "openai>=2.17.0",
    "langchain-openai>=1.1.8",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "pyspark>=4.1.1",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6; platform_system == \"Windows\"",
    "distro==1.9.0",
    "et-xmlfile==2.0.0",
    "greenlet==3.3.1; platform_machine == \"win32\" or platform_machine == \"WIN32\" or platform_machine == \"AMD64\" or platform_machine == \"amd64\" or platform_machine == \"x86_64\" or platform_machine == \"ppc64le\" or platform_machine == \"aarch64\"",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jiter==0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.9",
    "langchain-core==1.2.9",
    "langchain-openai==1.1.8",
    "langgraph==1.0.8",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.7",
    "langgraph-sdk==0.3.4",
    "langsmith==0.6.9",
    "numpy>=2.4.2",
    "openai==2.17.0",
    "openpyxl==3.1.5",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pandas==3.0.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyodbc==5.3.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "regex==2026.1.15",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.46",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3; sys_platform == \"emscripten\" or sys_platform == \"win32\"",
    "urllib3==2.6.3",
    "uuid-utils==0.14.0",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
    "langchain-anthropic>=1.3.2",
    "anthropic>=0.84.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "faiss-cpu>=1.13.2",
    "sentence-transformers>=5.2.3",
    "pypdf>=6.7.5",
    "langchain-huggingface>=1.2.0",
    "faiss-cpu",
    "sentence-transformers",
    "mcp[cli]>=1.26.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
dev = "python ./dev/server.py"
dev_ui = "python ./dev/server_ui.py"

test_tool1 = "python ./tests/test_client_tool1.py"
test_tool2 = "python ./tests/test_client_tool2.py"
test_ui = "python ./tests/test_client_pricing_debug.py"